refactor: data: Move attributes/__init__.py, property.py to models/common
