+++ alma_sbom/data/__init__.py
@@ -1,3 +1,3 @@
-from .models import Package, NullPackage, Build, PackageNevra, Iso
 from .collectors import ImmudbCollector, AlbsCollector, RpmCollector, IsoCollector
-from .attributes import Property
+from .models import Package, NullPackage, Build, PackageNevra, Iso
+from .models.common import Property
