+++ alma_sbom/data/models/__init__.py
@@ -1,3 +1,15 @@
+### Core Data Model
 from .package import Package, NullPackage, PackageNevra
 from .build import Build
 from .iso import Iso
+
+### Common Data Model
+from .common import (
+    PackageProperties,
+    BuildSourceProperties,
+    GitSourceProperties,
+    SrpmSourceProperties,
+    BuildPropertiesForPackage,
+    BuildPropertiesForBuild,
+    SBOMProperties,
+)
