+++ alma_sbom/data/models/common/property.py
